requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
requests==2.22.0
urllib3==1.25.7
apache-airflow>=1.9.0
apache-beam>=2.4.0
tzlocal>=1.5.1
